rag-mvp/backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
numpy==2.1.3
tqdm==4.66.4
fastembed==0.7.1
pypdf==4.2.0
python-docx==1.1.2
markdown==3.6
requests==2.32.3

